% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_fit.R
\name{fit_gradation}
\alias{fit_gradation}
\title{Gradation Curve Fitting}
\usage{
fit_gradation(gradation_tbl, sizes, method = "SP")
}
\arguments{
\item{gradation_tbl}{Gradation table (data frame).}

\item{sizes}{The sizes to interpolate the \% passing.}

\item{method}{One of "GS" (Gaudin-Schuhmann), "RR" (Rosin-Rammler), "SB" (Swebrec) or splines otherwise.}
}
\value{
A vector of passing through percentages.
}
\description{
Gradation Curve Fitting
}
