% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_fit.R
\name{gaudin_schuhmann}
\alias{gaudin_schuhmann}
\title{Gaudin-Schuhmann curve}
\usage{
gaudin_schuhmann(x, k, m)
}
\arguments{
\item{x}{A numeric vector with the sizes.}

\item{k}{Gradation topsize, in the same units as \code{x}.}

\item{m}{An exponent.}
}
\value{
A vector of passing through percentages.
}
\description{
Gaudin-Schuhmann curve
}
