% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_fit.R
\name{rosin_rammler}
\alias{rosin_rammler}
\title{Rosin-Rammler curve}
\usage{
rosin_rammler(x, a, b)
}
\arguments{
\item{x}{A numeric vector with the sizes.}

\item{a}{The size in which 63.2\% of the feed passes through.}

\item{b}{An exponent.}
}
\value{
A vector of passing through percentages.
}
\description{
Rosin-Rammler curve
}
