% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_distributions.R
\docType{data}
\name{size_distributions}
\alias{size_distributions}
\title{Size distributions (gradations) of 22 streams.}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{id}{stream description}
\item{x}{sieve size, in mm}
\item{y}{\% passing through}
\item{stream}{measurement origin}
...
}
}
\source{
\url{www.linkedin.com/in/odneto}
}
\usage{
size_distributions
}
\description{
A dataset containing sizes and \% passing - the gradation -  of 22 crushed stone streams.
}
\keyword{datasets}
