% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_fit.R
\name{swebrec}
\alias{swebrec}
\title{The Swebrec curve}
\usage{
swebrec(x, b, xmax, x50)
}
\arguments{
\item{x}{A numeric vector with the sizes.}

\item{b}{An exponent.}

\item{xmax}{Gradation topsize, in the same units as \code{x}.}

\item{x50}{Gradation P50, in the same units as \code{x}.}
}
\value{
A vector of passing through percentages.
}
\description{
The Swebrec curve
}
